MOBSDK-662: SDK: Ensure 64-bit compatibility -- Note, due to an API incompatibility this update does not take into account any changes required for ObjectiveCMIS-64bit

+++ AlfrescoSDK/AlfrescoSDK/Services/AlfrescoDocumentFolderService.m
@@ -546,7 +546,7 @@ typedef void (^CMISObjectCompletionBlock)(CMISObject *cmisObject, NSError *error
                     {
                         sortedChildren = @[];
                     }
-                    pagingResult = [[AlfrescoPagingResult alloc] initWithArray:sortedChildren hasMoreItems:pagedResult.hasMoreItems totalItems:pagedResult.numItems];
+                    pagingResult = [[AlfrescoPagingResult alloc] initWithArray:sortedChildren hasMoreItems:pagedResult.hasMoreItems totalItems:(int)pagedResult.numItems];
 //                    pagingResult = [AlfrescoPagingUtils pagedResultFromArray:sortedChildren listingContext:listingContext];
                     completionBlock(pagingResult, nil);
                 }
